Moving _pb_timestamp_to_datetime into core.
This is in advance of `v1beta3`, where it will be needed
to parse `Value.timestamp_value` (which is of type
`timestamp_pb2.Timestamp`). Also adding `_datetime_to_pb_timestamp`
for the other direction.
